Remove slug support on Erlef.Community.Event (#351)
* Remove slug support on Erlef.Community.Event

We don't currently utilize slugs on events at this point and slugs end up blocking certain recurring events
from being added without coming up with a clever title in order to generate a unique slug. If for some reason
this changes in the future, we can simply add it back.