Blog: add posts, tags, versions for blogs
This commit adds the initial database schemas for blog posts,
tags and versions. Versions are meant to be embedded into posts,
hence no separate migration for them.

This commit does not use or alter the existing blog post schema,
because we want to be able to deploy changes to production without
affecting how the existing site functions until we are ready to
switch over to the new system.