Handle the case when ics description folding is broken
When we pull in ics files via xhr in calendar.js line folding
for long descriptions get managled. Specifically, rfc 2445 states
that lines that fold on a description line must be preceeded with a space.

This commit handles a very specific case when the there is no space on folded lines
that begins with an escape (e.g, \\,).